backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-socketio==5.12.1
aiohttp==3.11.11
python-multipart==0.0.20
pydantic-settings==2.7.1
python-dotenv==1.0.1